extract `range_except` and `range_diff` out of `buffer.rs`
